// decision-making-tool/src/app/enums.ts
export enum wheelColors {
    DARK = 'black',
    COLOR_STROKE = '#6a70a4',
    LIGHT = 'white',
}

export enum buttonName {
    START = 'Start',
    ADD_OPTION = 'Add Option',
    UPLOAD_FILE = 'Upload option from File',
    SAVE_FILE = 'Save options to File',
    CLEAR_ALL = 'Clear All Options',
    PASTE_LIST = 'Paste List',
    DELETE = 'Delete',
    BACK = 'Back',
    PICK = 'Pick',
    BACK_HOME = 'Back to main page',
    CONFIRM = 'Confirm',
    CANCEL = 'Cancel',
}

export enum inputPlaceholders {
    OPTION_TITLE = 'Option Title',
    WEIGHT = 'Weight',
    PASTE_TEXTAREA = 'Enter data in CSV format: title, weight',
}

export enum pagePath {
    OPTIONS = 'options',
    PICKER = 'decision-picker',
    ERROR = 'error',
    DEFAULT = '',
}